HADOOP-4759. Removes temporary output directory for failed and killed tasks in the JobTracker's task commit thread. Contributed by Amareshwari Sriramadasu.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/core/trunk@741192 13f79535-47bb-0310-9956-ffa450edef68

+++ src/mapred/org/apache/hadoop/mapreduce/lib/output/FileOutputCommitter.java
@@ -174,8 +174,10 @@ public class FileOutputCommitter extends OutputCommitter {
   @Override
   public void abortTask(TaskAttemptContext context) {
     try {
-      context.progress();
-      outputFileSystem.delete(workPath, true);
+      if (workPath != null) { 
+        context.progress();
+        outputFileSystem.delete(workPath, true);
+      }
     } catch (IOException ie) {
       LOG.warn("Error discarding output" + StringUtils.stringifyException(ie));
     }
